Refactored naive trend models
