+++ gtime/forecasting/tests/test_trend.py
