+++ gtime/forecasting/trend.py
